now expression parser support float

+++ Expression_Parser/expression_parser.h
@@ -27,7 +27,7 @@
 #include <stdint.h>
 
 uint8_t Validate_Expression(const char *str);
-int32_t Evaluate_Expression(const char *str);
-uint8_t Evaluate_Expression2(const char *str, int32_t *value, uint8_t *sign);
+float Evaluate_Expression(const char *str);
+uint8_t Evaluate_Expression2(const char *str, float *value, uint8_t *sign);
 
 #endif /* EXPRESSION_PARSER_H_ */
